+++ dandy/llm/service/prompts.py
@@ -8,25 +8,13 @@ from dandy.intel import BaseIntel
 from dandy.llm.prompt import Prompt
 
 
-def service_system_intel_prompt(
-        intel_class: Type[BaseIntel],
+def service_system_prompt(
         prefix_system_prompt: Union[Prompt, None] = None
 ) -> Prompt:
     prompt = Prompt()
 
     if isinstance(prefix_system_prompt, Prompt):
         prompt.prompt(prefix_system_prompt)
-
-    prompt.line_break()
-    prompt.text('You must use the following rules when responding to the user.')
-    prompt.ordered_list([
-        'Your response must be valid JSON.',
-        'Your response must be valid according to the JSON schema provided below.',
-        'You will always respond with a JSON object with some values.',
-    ])
-    prompt.line_break()
-    prompt.text('JSON Schema:')
-    prompt.intel_schema(intel_class, triple_quote=True)
 
     return prompt