+++ dandy/llm/service/request/openai.py
@@ -3,13 +3,9 @@ from typing_extensions import Union
 from dandy.llm.service.request.request import BaseRequestBody
 
 
-_JSON_RESPONSE_FORMAT = {'type': 'json_object'}
-_TEXT_RESPONSE_FORMAT = {'type': 'text'}
-
-
 class OpenaiRequestBody(BaseRequestBody):
     stream: bool = False
-    response_format: dict = _JSON_RESPONSE_FORMAT
+    response_format: dict = { type: 'json_schema', 'json_schema': {'strict': True, 'schema': ...} }
     max_completion_tokens: Union[int, None] = None
     seed: Union[int, None] = None
     temperature: Union[float, None] = None
@@ -26,8 +22,8 @@ class OpenaiRequestBody(BaseRequestBody):
     def get_temperature(self) -> float:
         return self.temperature
 
-    def set_format_to_json(self):
-        self.response_format = _JSON_RESPONSE_FORMAT
+    def set_format_to_json_schema(self, json_schema: dict):
+        self.response_format['json_schema']['schema'] = json_schema
 
     def set_format_to_text(self):
-        self.response_format = _TEXT_RESPONSE_FORMAT
+        self.response_format = {'type': 'text'}