+++ tests/llm/test_utils.py
@@ -3,8 +3,6 @@ from unittest import TestCase
 from dandy.llm import Prompt
 from dandy.llm.utils import get_estimated_token_count_for_prompt
 
-from example.pirate.ship.intelligence.intel import ShipIntel
-
 
 class TestUtils(TestCase):
     def test_get_estimated_token_count_for_prompt(self):
@@ -12,8 +10,7 @@ class TestUtils(TestCase):
             prompt=(
                 Prompt()
                 .text('Hello World')
-            ),
-            intel_class=ShipIntel,
+            )
         )
 
         self.assertGreater(token_count, 0)